requirements.txt
cryptography==42.0.5
substrate-interface==1.7.7
pysha3==1.0.2
